pythonista/requirements.txt
appdirs==1.4.3
apturl==0.5.2
bcrypt==3.1.7
beautifulsoup4==4.9.3
binwalk==2.2.0
blinker==1.4
brainfuck-interpreter==1.0.0
Brlapi==0.7.0
bs4==0.0.1
capstone==4.0.2
certifi==2019.11.28
chardet==3.0.4
Click==7.0
colorama==0.4.3
command-not-found==0.3
cryptography==2.8
cupshelpers==1.0
cycler==0.10.0
dbus-python==1.2.16
decorator==4.4.2
defer==1.0.6
distlib==0.3.0
distro==1.4.0
distro-info===0.23ubuntu1
duplicity==0.8.12.0
entrypoints==0.3
evdev==1.4.0
fasteners==0.14.1
filelock==3.0.12
future==0.18.2
gTTS==2.2.2
hashID==3.1.4
httplib2==0.14.0
idna==2.8
importlib-metadata==1.5.0
intervaltree==3.1.0
keyboard==0.13.5
keyring==18.0.1
kiwisolver==1.0.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lockfile==0.12.2
louis==3.12.0
lxml==4.6.3
macaroonbakery==1.3.1
Mako==1.1.0
MarkupSafe==1.1.0
matplotlib==3.1.2
monotonic==1.5
more-itertools==4.2.0
mpmath==1.2.1
netifaces==0.10.4
numpy==1.17.4
oauthlib==3.1.0
olefile==0.46
packaging==20.9
paramiko==2.6.0
pbkdf2==1.3
pexpect==4.6.0
Pillow==7.0.0
playsound==1.2.2
protobuf==3.6.1
psutil==5.8.0
pwntools==4.3.1
pyaes==1.6.1
pycairo==1.16.2
pycryptodome==3.4.3
pycryptodome-test-vectors==1.0.2
pycryptodomex==3.10.1
pycups==1.9.73
pyelftools==0.27
pygame==2.0.1
Pygments==2.8.1
PyGObject==3.36.0
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
pynput==1.7.3
PyOpenGL==3.1.0
pyparsing==2.4.7
PyQt5==5.14.1
pyqtgraph==0.11.0rc0
pyRFC3339==1.1
pyserial==3.5
PySocks==1.7.1
python-apt==2.0.0+ubuntu0.20.4.4
python-brainfuck==0.9.1
python-dateutil==2.7.3
python-debian===0.1.36ubuntu1
python-xlib==0.29
pytz==2019.3
pyxdg==0.26
pyxhook==1.0.0
PyYAML==5.3.1
reportlab==3.5.34
requests==2.22.0
requests-futures==1.0.0
requests-unixsocket==0.2.0
ROPGadget==6.5
scipy==1.3.3
SecretStorage==2.3.1
simplejson==3.16.0
sip==4.19.21
six==1.14.0
sortedcontainers==2.3.0
soupsieve==2.2.1
stem==1.8.0
sympy==1.7.1
systemd-python==234
torrequest==0.1.0
ubuntu-advantage-tools==20.3
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
unicorn==1.0.2rc3
urllib3==1.25.8
usb-creator==0.3.7
virtualenv==20.0.17
wadllib==1.3.3
xkit==0.0.0
zipp==1.0.0
